Ported spatial point and distance function to 3.0
