--- community/cypher/frontend-2.3/src/main/scala/org/neo4j/cypher/internal/frontend/v2_3/symbols/GeometryType.scala
@@ -1,10 +1,0 @@
-package org.neo4j.cypher.internal.frontend.v2_3.symbols
-
-object GeometryType {
-  val instance = new GeometryType() {
-    val parentType = CTAny
-    override val toString = "Geometry"
-  }
-}
-
-sealed abstract class GeometryType extends CypherType
